split the 2000 line v80.c into 41 much smaller files
- polyfill/ a new directory with replacements for likely candidates
  for missing system headers and apis.  Note: it's not a replacement
  for the system library, only fallbacks for apis used by this
  project
- error.c: error handling
- file.h, file.c: file handling
- stack.c: simple generic stack datatype
- hash.c: simple hash table datatype with buckets made with stack.c
- symtab.c: symbols and a symbol table for them made with hash.c
- token.h, token.c: token data type, and a line at a time tokenizer
- parser.c: recursive descent parser for v80 assembly using token.c
- main.c: command line processing, and driver for feeding the parser
- Makefile: simple rules for making versions of v80 from the above to
  check that it works with almost nothing from libc using c89, and also
  using c99 with optimized libc functions instead of polyfills.
- README.md: A little about how to build and use it all.
