+++ bootstrap/error.c
@@ -1,0 +1,109 @@
+#ifndef SEEN_V80_ERROR_C
+#define SEEN_V80_ERROR_C
+
+#include "polyfill/stdio.h"
+#include "polyfill/stdlib.h"
+#include "polyfill/string.h"
+
+#include "file.h"
+
+/* ERROR HANDLING */
+
+/* After this point, all remaining code calls err_fatal() with an error code to
+   bail out with a diagnostic that includes the file name, position and an error
+   message, along with context about the current token if provided. */
+
+#define NOTHING_ELSE
+#define STRINGIFY(_x)       #_x
+#define XSTRINGIFY(_x)      STRINGIFY(_x)
+
+#define ERRORHANDLING                                                           \
+    X(ERR_BADBYTE,          "Value overflows 8-bits")                           \
+    X(ERR_BADBYTEEXPRESSION,"Expression overflows 8-bits")                      \
+    X(ERR_BADCHAR,          "Invalid char literal")                             \
+    X(ERR_BADCONDITION,     "Invalid condition")                                \
+    X(ERR_BADDIGIT,         "Invalid number digit")                             \
+    X(ERR_BADEXPRESSION,    "Invalid expression")                               \
+    X(ERR_BADFILE,          "File not found")                                   \
+    X(ERR_BADFILENAME,      "Invalid filename")                                 \
+    X(ERR_BADINSTRUCTION,   "Invalid instruction")                              \
+    X(ERR_BADKEYWORD,       "Invalid keyword")                                  \
+    X(ERR_BADLABEL,         "Label exceeds " XSTRINGIFY(TOKEN_MAXLEN) " chars") \
+    X(ERR_BADNUMBER,        "Invalid number")                                   \
+    X(ERR_BADSTRING,        "Invalid string")                                   \
+    X(ERR_BADSYMBOL,        "Symbol exceeds " XSTRINGIFY(TOKEN_MAXLEN) " chars")\
+    X(ERR_BADVALUE,         "Invalid value")                                    \
+    X(ERR_DUPLABEL,         "Labels cannot be redefined")                       \
+    X(ERR_EXPECTEXPRESSION, "Expected expression")                              \
+    X(ERR_EXPECTFILENAME,   "Expected filename")                                \
+    X(ERR_EXPECTMACRONAME,  "Expected macro name")                              \
+    X(ERR_EXPECTMACROBODY,  "Expected macro body")                              \
+    X(ERR_LINETOOLONG,      "Line exceeds " XSTRINGIFY(LINE_MAXLEN) " columns") \
+    X(ERR_LOCALORPHAN,      "Local label without preceding label")              \
+    X(ERR_PATHTOOLONG,      "Include path exceeds " XSTRINGIFY(MAXPATHLEN) " characters") \
+    X(ERR_NUMBERTOOBIG,     "Expression overflow")                              \
+    X(ERR_NOBRACKET,        "Invalid expression, missing ']'")                  \
+    X(ERR_NOFORWARDREF,     "Forward-references to labels not allowed here!")   \
+    X(ERR_NOPAREN,          "Invalid expression, missing ')'")                  \
+    X(ERR_NOSTRINGWORD,     "Invalid string argument to keyword")               \
+    X(ERR_UNDEFCONSTANT,    "Undefined constant reference")                     \
+    X(ERR_UNDEFLABEL,       "Undefined label reference")                        \
+NOTHING_ELSE
+
+enum ErrCode {
+#define X(_c, _m) _c,
+    ERRORHANDLING
+#undef X
+    ERR_NUMBEROFENTRIES
+};
+
+void
+err_print(FILE *stream, const char *err, const char *msg, unsigned len)
+{
+    fprintf(stream, "%s", err);
+    if(msg) {
+        const char *after = msg + (len ? len : strlen(msg));
+        fprintf(stream, " near '");
+        for(; msg < after; ++msg) {
+            if(*msg < (char)0x20)
+                fprintf(stream, "^%c", *msg + 0x40);
+            else if(*msg < (char)0x7f)
+                putc(*msg, stream);
+            else if(*msg == (char)0x7f)
+                fprintf(stream, "^?");
+            else if(*msg < (char)0xa0)
+                fprintf(stream, "^[^%c", *msg - 0x40);
+            else if(*msg < (char)0xff)
+                fprintf(stream, "^[%c", *msg - 0x80);
+            else
+                fprintf(stream, "^[^?");
+        }
+        putc('\'', stream);
+    }
+}
+
+__attribute__((noreturn)) void
+err_fatal(enum ErrCode code, const char *msg, unsigned len, unsigned col)
+{
+    fprintf(stderr, "v80: %s:%d", files->zfname, files->lineno);
+    if(col)
+        fprintf(stderr, ".%d", col);
+    fprintf(stderr, ": ERROR: ");
+    switch (code) {
+#define X(_c, _m) case _c: err_print(stderr, (_m), msg, len); break;
+        ERRORHANDLING
+#undef X
+        case ERR_NUMBEROFENTRIES:
+            fprintf(stderr, "internal error"); break;
+    }
+    putc('\n', stderr);
+    exit(EXIT_FAILURE);
+}
+
+static inline void
+err_fatal_str(enum ErrCode code, const char *str)
+{
+    err_fatal(code, str, 0, 0);
+}
+
+#endif
