+++ bootstrap/polyfill/limits.h
@@ -1,0 +1,8 @@
+#ifndef SEEN_POLYFILL_LIMITS_H
+#define SEEN_POLYFILL_LIMITS_H
+
+#ifndef UINT_MAX
+#  define UINT_MAX          ~0
+#endif
+
+#endif
