+++ bootstrap/polyfill/stdlib/xmalloc.c
@@ -1,0 +1,5 @@
+static inline void *
+xmalloc(unsigned nbytes)
+{
+    return xhavemem(malloc(nbytes));
+}
