+++ bootstrap/polyfill/string/strdup.c
@@ -1,0 +1,7 @@
+static inline char *
+zstrdup(const char *zsrc)
+{
+    return strndup(zsrc, strlen(zsrc));
+}
+#undef strdup
+#define strdup zstrdup
