+++ bootstrap/polyfill/string/strneq.c
@@ -1,0 +1,13 @@
+#include "polyfill/assert.h"
+
+static int
+zstrneq(const char *zp, const char *q, unsigned qlen)
+{
+    assert(zp && q);
+    /* compare `qlen` chars, or until we find a \0 or a mismatch */
+    while(*zp && qlen > 0 && *zp == *q)
+        ++zp, ++q, --qlen;
+    return (!*zp && !qlen) || (*zp == *q);
+}
+#undef strneq
+#define strneq zstrneq
