+++ bootstrap/polyfill/sys/stat/s_isreg.c
@@ -1,0 +1,12 @@
+static inline int
+s_isreg(const char *kpath)
+{
+#ifndef NO_SYS_STAT_H
+    struct stat statbuf;
+    /* If we have stat(2), diagnose attempt to read from anything but
+       a regular file. */
+    if(lstat(kpath, &statbuf) == 0)
+        return(S_ISREG(statbuf.st_mode));
+#endif
+    return 1;
+}
